+++ Terraform/DemoFullStack-Modularized/modules/fsx-ontap/main.tf
@@ -379,12 +379,14 @@ resource "aws_ssm_document" "ssm_fsx_ontap_setup" {
                               Exit 1
                           }
 
-                          Write-Output 'Removing CredSSP registry entries'
-                          Try {
-                              Remove-Item -Path (Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey) -Force -Recurse -ErrorAction Stop
-                          } Catch [System.Exception] {
-                              Write-Output "Failed to remove CredSSP registry entries $_"
-                              Exit 1
+                          If (Test-Path -Path $(Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey)) {
+                              Write-Output 'Removing CredSSP registry entries'
+                              Try {
+                                  Remove-Item -Path (Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey) -Force -Recurse -ErrorAction Stop
+                              } Catch [System.Exception] {
+                                  Write-Output "Failed to remove CredSSP registry entries $_"
+                                  #Exit 1
+                              }
                           }
                       }
                       Default { 
@@ -836,12 +838,14 @@ resource "aws_ssm_document" "ssm_fsx_ontap_alias" {
                               Exit 1
                           }
 
-                          Write-Output 'Removing CredSSP registry entries'
-                          Try {
-                              Remove-Item -Path (Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey) -Force -Recurse -ErrorAction Stop
-                          } Catch [System.Exception] {
-                              Write-Output "Failed to remove CredSSP registry entries $_"
-                              Exit 1
+                          If (Test-Path -Path $(Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey)) {
+                              Write-Output 'Removing CredSSP registry entries'
+                              Try {
+                                  Remove-Item -Path (Join-Path -Path "Registry::$RootKey" -ChildPath $CredDelKey) -Force -Recurse -ErrorAction Stop
+                              } Catch [System.Exception] {
+                                  Write-Output "Failed to remove CredSSP registry entries $_"
+                                  #Exit 1
+                              }
                           }
                       }
                       Default {
